requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.18.0
pydantic>=1.8.0
python-dotenv>=0.19.0
pytest>=6.2.0
pytest-asyncio>=0.15.0
